--- .eslintrc
@@ -1,24 +1,0 @@
-{
-  "env": {
-    "jquery": true,
-    "browser": true,
-    "es2020": true
-  },
-  "root": true,
-  "parser": "@typescript-eslint/parser",
-  "parserOptions": {
-    "ecmaVersion": 2020,
-    "sourceType": "module"
-  },
-  "extends": ["plugin:@typescript-eslint/recommended", "@typhonjs-fvtt/eslint-config-foundry.js"],
-  "rules": {
-    "no-underscore-dangle": "off",
-    "import/extensions": "off",
-    "class-methods-use-this": [
-      "error",
-      {
-        "exceptMethods": ["getData", "_updateObject"]
-      }
-    ]
-  }
-}
