--- .github/workflows/get-version.js
@@ -1,2 +1,0 @@
-var fs = require('fs');
-console.log(JSON.parse(fs.readFileSync('./dist/module.json', 'utf8')).version);
