+++ .github/workflows/main.yml
@@ -1,15 +1,19 @@
 name: Module CI/CD
 
 on:
-  push:
-    tags:
-      - 'v*'
+  release:
+    types: [published]
 
 jobs:
   build:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v2
+
+      # get part of the tag after the `v`
+      - name: Extract tag version number
+        id: get_version
+        uses: battila7/get-version-action@v2
 
       - name: Use Node.js
         uses: actions/setup-node@v1
@@ -20,34 +24,38 @@ jobs:
         run: npm ci
 
       # run our gulp build action
-      - name: Build dist directory
+      - name: Build CSS File
         run: npm run build
+
+      # Substitute the Manifest and Download URLs in the module.json
+      - name: Substitute Manifest and Download Links For Versioned Ones
+        id: sub_manifest_link_version
+        uses: microsoft/variable-substitution@v1
+        with:
+          files: 'module.json'
+        env:
+          version: ${{steps.get_version.outputs.version-without-v}}
+          url: https://github.com/${{github.repository}}
+          manifest: https://github.com/${{github.repository}}/releases/latest/download/module.json
+          download: https://github.com/${{github.repository}}/releases/download/${{github.event.release.tag_name}}/module.zip
 
       # create a zip file with all files required by the module to add to the release
       - name: Zip Files
-        working-directory: ./dist
-        run: zip -r ./module.zip ./*
-
-      # Get the version from 'module.json'
-      - name: Get Version
-        shell: bash
-        id: get-version
-        run: echo "::set-output name=version::$(node ./.github/workflows/get-version.js)"
+        run: zip -r ./module.zip module.json LICENSE foundryvtt-devMode.mjs foundryvtt-devMode.css module/ templates/ lang/
 
       # Create a release for this specific version
-      - name: Create Release
+      - name: Update Release with Files
         id: create_version_release
         uses: ncipollo/release-action@v1
         with:
-          allowUpdates: true # set this to false if you want to prevent updating existing releases
-          name: ${{ steps.get-version.outputs.version }}
-          body: |
-            ${{ steps.changelog.outputs.clean_changelog }}
-          draft: false
-          prerelease: false
+          allowUpdates: true # Set this to false if you want to prevent updating existing releases
+          name: ${{ github.event.release.name }}
+          draft: ${{ github.event.release.unpublished }}
+          prerelease: ${{ github.event.release.prerelease }}
           token: ${{ secrets.GITHUB_TOKEN }}
-          artifacts: './dist/module.json,./dist/module.zip'
-          tag_name: ${{ steps.get-version.outputs.version }}
+          artifacts: './module.json, ./module.zip'
+          tag: ${{ github.event.release.tag_name }}
+          body: ${{ github.event.release.body }}
 
       # Publish this new version to the Foundry VTT Module Listing
       - name: FoundryVTT AutoPublish
@@ -57,4 +65,4 @@ jobs:
           password: ${{ secrets.FOUNDRY_ADMIN_PW }}
           module-id: 963
           manifest-url: https://github.com/${{ github.repository }}/releases/download/${{ steps.get-version.outputs.version }}/module.json
-          manifest-file: src/module.json
+          manifest-file: module.json
