+++ .gitignore
@@ -1,3 +1,22 @@
-node_modules
-foundryconfig.json
-dist
+# IDE
+.idea/
+.vs/
+
+# Node Modules
+node_modules/
+npm-debug.log
+
+# yarn2
+.yarn/*
+!.yarn/patches
+!.yarn/releases
+!.yarn/plugins
+!.yarn/sdks
+!.yarn/versions
+.pnp.*
+
+# Distribution files
+**/*.css
+
+# ESLint
+.eslintcache
