--- .prettierignore
@@ -1,2 +1,0 @@
-# Ignore all HBS files
-*.hbs
