+++ .prettierrc
@@ -1,8 +1,7 @@
 {
   "semi": true,
-  "trailingComma": "es5",
+  "trailingComma": "all",
   "singleQuote": true,
   "printWidth": 120,
-  "tabWidth": 2,
-  "endOfLine": "auto"
+  "tabWidth": 2
 }
