+++ docs/debug-mode-core-config.png
