+++ docs/debug-mode-package-config.png
