+++ package.json
@@ -2,14 +2,42 @@
   "private": true,
   "name": "foundryvtt-devMode",
   "version": "1.3.0",
-  "description": "",
+  "description": "A module with some tools to help enable Foundry VTT package developers.",
+  "license": "MIT",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/League-of-Foundry-Developers/foundryvtt-devMode"
+  },
+  "bugs": {
+    "url": "https://github.com/League-of-Foundry-Developers/foundryvtt-devMode/issues"
+  },
   "scripts": {
-    "publish": "gulp publish",
-    "lint": "prettier --write .",
     "build": "gulp build",
     "build:watch": "gulp watch",
-    "clean": "gulp clean"
+    "clean": "gulp clean",
+    "lint": "eslint --ext .js .",
+    "lint:fix": "eslint --ext .js --fix .",
+    "format": "prettier --write \"./**/*.(js|json|scss)\"",
+    "prepare": "husky install"
   },
-  "author": "",
-  "license": ""
+  "devDependencies": {
+    "@typhonjs-fvtt/eslint-config-foundry.js": "^0.8.0",
+    "babel-eslint": "^10.1.0",
+    "chalk": "^4.1.2",
+    "eslint": "^7.32.0",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-prettier": "^3.4.0",
+    "fs-extra": "^10.0.0",
+    "gulp": "^4.0.2",
+    "gulp-dart-sass": "^1.0.2",
+    "husky": "^7.0.1",
+    "lint-staged": "^11.1.2",
+    "prettier": "^2.3.2",
+    "sass": "^1.37.5"
+  },
+  "lint-staged": {
+    "*.(js)": "eslint --fix",
+    "*.(json|scss)": "prettier --write"
+  },
+  "dependencies": {}
 }
