--- src/module.json
@@ -1,40 +1,0 @@
-{
-	"name": "_dev-mode",
-	"title": "🧙 Developer Mode",
-	"description": "Enables developers to do developer things without affecting non-developers.",
-	"version": "1.3.0",
-	"author": "Andrew Krigline (akrigline)",
-	"authors": [
-		{
-			"name": "Andrew Krigline",
-			"discord": "Calego#0914",
-			"url": "https://github.com/ElfFriend-DnD",
-			"patreon": "https://www.patreon.com/ElfFriend_DnD"
-		}
-	],
-	"esmodules": [
-		"foundryvtt-devMode.js"
-	],
-	"styles": [
-		"foundryvtt-devMode.css"
-	],
-	"languages": [
-		{
-			"lang": "en",
-			"name": "English",
-			"path": "lang/en.json"
-		}
-	],
-	"minimumCoreVersion": "0.7.9",
-	"compatibleCoreVersion": "0.8.8",
-	"bugs": "https://github.com/League-of-Foundry-Developers/foundryvtt-devMode/issues",
-	"changelog": "https://github.com/League-of-Foundry-Developers/foundryvtt-devMode/releases",
-	"download": "https://github.com/League-of-Foundry-Developers/foundryvtt-devMode/releases/download/v1.3.0/module.zip",
-	"manifest": "https://github.com/League-of-Foundry-Developers/foundryvtt-devMode/releases/latest/download/module.json",
-	"readme": "https://raw.githubusercontent.com/League-of-Foundry-Developers/foundryvtt-devMode/main/README.md",
-	"url": "https://github.com/League-of-Foundry-Developers/foundryvtt-devMode",
-	"library": true,
-	"flags": {
-		"allowBugReporter": true
-	}
-}
